Move consumer
